product-service/modules/catalog/presentation/serializers/brand_serializer.py
from rest_framework import serializers

from ...infrastructure.models import Brand


class BrandSerializer(serializers.ModelSerializer):
    class Meta:
        model = Brand
        fields = ['id', 'slug', 'name', 'country']
