product-service/requirements.txt
Django==5.2.12
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
psycopg2-binary==2.9.9
cryptography==41.0.7
python-dotenv==1.0.0
